requirements.txt
google-cloud-speech>=2.21.0
sounddevice>=0.4.6
numpy>=1.24.3
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
openai>=1.3.0
noisereduce>=2.0.0
scipy>=1.11.0
